Fix instance switching

+++ src/main/java/net/minestom/server/network/packet/server/play/RespawnPacket.java
@@ -18,7 +18,7 @@ public class RespawnPacket implements ServerPacket {
 
     @Override
     public void write(@NotNull BinaryWriter writer) {
-        writer.writeByte(dimensionType.getId());
+        writer.writeInt(dimensionType.getId());
         writer.writeByte(difficulty.getId());
         writer.writeByte(gameMode.getId());
         writer.writeSizedString(levelType.getId());
